Decouple REST resources from primitive types and load arbitrary resources from the classpath.
