+++ core/src/main/java/io/atomix/core/set/DistributedSetType.java
@@ -20,7 +20,6 @@ import io.atomix.core.collection.impl.CollectionUpdateResult;
 import io.atomix.core.iterator.impl.IteratorBatch;
 import io.atomix.core.set.impl.DefaultDistributedSetBuilder;
 import io.atomix.core.set.impl.DefaultDistributedSetService;
-import io.atomix.core.set.impl.DistributedSetResource;
 import io.atomix.core.set.impl.SetUpdate;
 import io.atomix.core.transaction.TransactionId;
 import io.atomix.core.transaction.TransactionLog;
@@ -83,11 +82,6 @@ public class DistributedSetType<E> implements PrimitiveType<DistributedSetBuilde
   @Override
   public PrimitiveService newService(ServiceConfig config) {
     return new DefaultDistributedSetService<>();
-  }
-
-  @Override
-  public Class<?> getResourceClass() {
-    return DistributedSetResource.class;
   }
 
   @Override
