+++ core/src/main/java/io/atomix/core/value/DistributedValueType.java
@@ -17,7 +17,6 @@ package io.atomix.core.value;
 
 import io.atomix.core.value.impl.DefaultDistributedValueBuilder;
 import io.atomix.core.value.impl.DefaultDistributedValueService;
-import io.atomix.core.value.impl.DistributedValueResource;
 import io.atomix.primitive.PrimitiveManagementService;
 import io.atomix.primitive.PrimitiveType;
 import io.atomix.primitive.service.PrimitiveService;
@@ -51,11 +50,6 @@ public class DistributedValueType<V> implements PrimitiveType<DistributedValueBu
   @Override
   public PrimitiveService newService(ServiceConfig config) {
     return new DefaultDistributedValueService();
-  }
-
-  @Override
-  public Class<?> getResourceClass() {
-    return DistributedValueResource.class;
   }
 
   @Override
