--- primitive/src/main/java/io/atomix/primitive/resource/PrimitiveResourceFactory.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2018-present Open Networking Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.atomix.primitive.resource;
-
-import io.atomix.primitive.DistributedPrimitive;
-
-/**
- * Primitive resource factory.
- */
-@FunctionalInterface
-public interface PrimitiveResourceFactory<P extends DistributedPrimitive> {
-
-  /**
-   * Creates a new primitive resource.
-   *
-   * @param primitive the primitive instance
-   * @return the resource instance
-   */
-  PrimitiveResource create(P primitive);
-
-}
