+++ rest/src/main/java/io/atomix/rest/AtomixResource.java
@@ -13,8 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.atomix.rest;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
- * Provides interfaces for defining REST APIs for distributed primitives.
+ * REST API resource.
  */
-package io.atomix.primitive.resource;
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface AtomixResource {
+}
