+++ rest/src/main/java/io/atomix/rest/impl/EventLog.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.utils;
+package io.atomix.rest.impl;
 
 import java.util.Queue;
 import java.util.concurrent.CompletableFuture;
