+++ rest/src/main/java/io/atomix/rest/impl/EventManager.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.utils;
+package io.atomix.rest.impl;
 
 import java.util.Map;
 import java.util.Set;
