+++ rest/src/main/java/io/atomix/rest/resources/AtomicMapResource.java
@@ -13,18 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.collection.impl;
+package io.atomix.rest.resources;
 
 import com.google.common.collect.Sets;
-import io.atomix.core.collection.AsyncDistributedCollection;
-import io.atomix.core.collection.DistributedCollectionConfig;
-import io.atomix.primitive.PrimitiveType;
-import io.atomix.primitive.resource.PrimitiveResource;
+import io.atomix.core.map.AsyncAtomicMap;
+import io.atomix.core.map.AtomicMapConfig;
+import io.atomix.core.map.AtomicMapType;
+import io.atomix.rest.AtomixResource;
+import io.atomix.utils.time.Versioned;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -35,40 +38,27 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 /**
- * Distributed collection resource.
+ * Atomic map resource.
  */
-public abstract class DistributedCollectionResource<P extends AsyncDistributedCollection<String>, C extends DistributedCollectionConfig<C>> extends PrimitiveResource<P, C> {
-  private final Logger LOGGER = LoggerFactory.getLogger(getClass());
+@AtomixResource
+@Path("/atomic-map")
+public class AtomicMapResource extends PrimitiveResource<AsyncAtomicMap<String, String>, AtomicMapConfig> {
+  private static final Logger LOGGER = LoggerFactory.getLogger(AtomicMapResource.class);
 
-  protected DistributedCollectionResource(PrimitiveType type) {
-    super(type);
+  public AtomicMapResource() {
+    super(AtomicMapType.instance());
   }
 
   @GET
-  @Path("/{name}")
+  @Path("/{name}/{key}")
   @Produces(MediaType.APPLICATION_JSON)
   public void get(
       @PathParam("name") String name,
+      @PathParam("key") String key,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).whenComplete((collection, error) -> {
+    getPrimitive(name).thenCompose(map -> map.get(key)).whenComplete((result, error) -> {
       if (error == null) {
-        response.resume(Response.ok(Sets.newHashSet(collection.iterator().sync())));
-      } else {
-        response.resume(Response.serverError().build());
-      }
-    });
-  }
-
-  @PUT
-  @Path("/{name}/{element}")
-  @Produces(MediaType.APPLICATION_JSON)
-  public void add(
-      @PathParam("name") String name,
-      @PathParam("element") String element,
-      @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(collection -> collection.add(element)).whenComplete((result, error) -> {
-      if (error == null) {
-        response.resume(Response.ok(result).build());
+        response.resume(Response.ok(result != null ? new VersionedResult(result) : null).build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
@@ -76,16 +66,18 @@ public abstract class DistributedCollectionResource<P extends AsyncDistributedCo
     });
   }
 
-  @GET
-  @Path("/{name}/{element}")
+  @PUT
+  @Path("/{name}/{key}")
+  @Consumes(MediaType.TEXT_PLAIN)
   @Produces(MediaType.APPLICATION_JSON)
-  public void contains(
+  public void put(
       @PathParam("name") String name,
-      @PathParam("element") String element,
+      @PathParam("key") String key,
+      String value,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(collection -> collection.contains(element)).whenComplete((result, error) -> {
+    getPrimitive(name).thenCompose(map -> map.put(key, value)).whenComplete((result, error) -> {
       if (error == null) {
-        response.resume(Response.ok(result).build());
+        response.resume(Response.ok(result != null ? new VersionedResult(result) : null).build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
@@ -94,15 +86,31 @@ public abstract class DistributedCollectionResource<P extends AsyncDistributedCo
   }
 
   @DELETE
-  @Path("/{name}/{element}")
+  @Path("/{name}/{key}")
   @Produces(MediaType.APPLICATION_JSON)
   public void remove(
       @PathParam("name") String name,
-      @PathParam("element") String element,
+      @PathParam("key") String key,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(collection -> collection.remove(element)).whenComplete((result, error) -> {
+    getPrimitive(name).thenCompose(map -> map.remove(key)).whenComplete((result, error) -> {
       if (error == null) {
-        response.resume(Response.ok(result).build());
+        response.resume(Response.ok(result != null ? new VersionedResult(result) : null).build());
+      } else {
+        LOGGER.warn("{}", error);
+        response.resume(Response.serverError().build());
+      }
+    });
+  }
+
+  @GET
+  @Path("/{name}/keys")
+  @Produces(MediaType.APPLICATION_JSON)
+  public void keys(
+      @PathParam("name") String name,
+      @Suspended AsyncResponse response) {
+    getPrimitive(name).whenComplete((map, error) -> {
+      if (error == null) {
+        response.resume(Response.ok(Sets.newHashSet(map.keySet().iterator().sync())).build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
@@ -116,7 +124,7 @@ public abstract class DistributedCollectionResource<P extends AsyncDistributedCo
   public void size(
       @PathParam("name") String name,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(collection -> collection.size()).whenComplete((result, error) -> {
+    getPrimitive(name).thenCompose(map -> map.size()).whenComplete((result, error) -> {
       if (error == null) {
         response.resume(Response.ok(result).build());
       } else {
@@ -126,18 +134,37 @@ public abstract class DistributedCollectionResource<P extends AsyncDistributedCo
     });
   }
 
-  @DELETE
-  @Path("/{name}")
+  @POST
+  @Path("/{name}/clear")
   public void clear(
       @PathParam("name") String name,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(collection -> collection.clear()).whenComplete((result, error) -> {
+    getPrimitive(name).thenCompose(map -> map.clear()).whenComplete((result, error) -> {
       if (error == null) {
-        response.resume(Response.ok().build());
+        response.resume(Response.noContent().build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
       }
     });
+  }
+
+  /**
+   * Versioned JSON result.
+   */
+  static class VersionedResult {
+    private final Versioned<String> value;
+
+    public VersionedResult(Versioned<String> value) {
+      this.value = value;
+    }
+
+    public String getValue() {
+      return value.value();
+    }
+
+    public long getVersion() {
+      return value.version();
+    }
   }
 }
