+++ rest/src/main/java/io/atomix/rest/resources/DistributedCollectionResource.java
@@ -13,21 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.map.impl;
+package io.atomix.rest.resources;
 
 import com.google.common.collect.Sets;
-import io.atomix.core.map.AsyncAtomicMap;
-import io.atomix.core.map.AtomicMapConfig;
-import io.atomix.core.map.AtomicMapType;
-import io.atomix.primitive.resource.PrimitiveResource;
-import io.atomix.utils.time.Versioned;
+import io.atomix.core.collection.AsyncDistributedCollection;
+import io.atomix.core.collection.DistributedCollectionConfig;
+import io.atomix.primitive.PrimitiveType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
-import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -38,26 +34,40 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 /**
- * Atomic map resource.
+ * Distributed collection resource.
  */
-@Path("/atomic-map")
-public class AtomicMapResource extends PrimitiveResource<AsyncAtomicMap<String, String>, AtomicMapConfig> {
-  private static final Logger LOGGER = LoggerFactory.getLogger(AtomicMapResource.class);
+public abstract class DistributedCollectionResource<P extends AsyncDistributedCollection<String>, C extends DistributedCollectionConfig<C>> extends PrimitiveResource<P, C> {
+  private final Logger LOGGER = LoggerFactory.getLogger(getClass());
 
-  public AtomicMapResource() {
-    super(AtomicMapType.instance());
+  protected DistributedCollectionResource(PrimitiveType type) {
+    super(type);
   }
 
   @GET
-  @Path("/{name}/{key}")
+  @Path("/{name}")
   @Produces(MediaType.APPLICATION_JSON)
   public void get(
       @PathParam("name") String name,
-      @PathParam("key") String key,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(map -> map.get(key)).whenComplete((result, error) -> {
+    getPrimitive(name).whenComplete((collection, error) -> {
       if (error == null) {
-        response.resume(Response.ok(result != null ? new VersionedResult(result) : null).build());
+        response.resume(Response.ok(Sets.newHashSet(collection.iterator().sync())));
+      } else {
+        response.resume(Response.serverError().build());
+      }
+    });
+  }
+
+  @PUT
+  @Path("/{name}/{element}")
+  @Produces(MediaType.APPLICATION_JSON)
+  public void add(
+      @PathParam("name") String name,
+      @PathParam("element") String element,
+      @Suspended AsyncResponse response) {
+    getPrimitive(name).thenCompose(collection -> collection.add(element)).whenComplete((result, error) -> {
+      if (error == null) {
+        response.resume(Response.ok(result).build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
@@ -65,18 +75,16 @@ public class AtomicMapResource extends PrimitiveResource<AsyncAtomicMap<String, 
     });
   }
 
-  @PUT
-  @Path("/{name}/{key}")
-  @Consumes(MediaType.TEXT_PLAIN)
+  @GET
+  @Path("/{name}/{element}")
   @Produces(MediaType.APPLICATION_JSON)
-  public void put(
+  public void contains(
       @PathParam("name") String name,
-      @PathParam("key") String key,
-      String value,
+      @PathParam("element") String element,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(map -> map.put(key, value)).whenComplete((result, error) -> {
+    getPrimitive(name).thenCompose(collection -> collection.contains(element)).whenComplete((result, error) -> {
       if (error == null) {
-        response.resume(Response.ok(result != null ? new VersionedResult(result) : null).build());
+        response.resume(Response.ok(result).build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
@@ -85,31 +93,15 @@ public class AtomicMapResource extends PrimitiveResource<AsyncAtomicMap<String, 
   }
 
   @DELETE
-  @Path("/{name}/{key}")
+  @Path("/{name}/{element}")
   @Produces(MediaType.APPLICATION_JSON)
   public void remove(
       @PathParam("name") String name,
-      @PathParam("key") String key,
+      @PathParam("element") String element,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(map -> map.remove(key)).whenComplete((result, error) -> {
+    getPrimitive(name).thenCompose(collection -> collection.remove(element)).whenComplete((result, error) -> {
       if (error == null) {
-        response.resume(Response.ok(result != null ? new VersionedResult(result) : null).build());
-      } else {
-        LOGGER.warn("{}", error);
-        response.resume(Response.serverError().build());
-      }
-    });
-  }
-
-  @GET
-  @Path("/{name}/keys")
-  @Produces(MediaType.APPLICATION_JSON)
-  public void keys(
-      @PathParam("name") String name,
-      @Suspended AsyncResponse response) {
-    getPrimitive(name).whenComplete((map, error) -> {
-      if (error == null) {
-        response.resume(Response.ok(Sets.newHashSet(map.keySet().iterator().sync())).build());
+        response.resume(Response.ok(result).build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
@@ -123,7 +115,7 @@ public class AtomicMapResource extends PrimitiveResource<AsyncAtomicMap<String, 
   public void size(
       @PathParam("name") String name,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(map -> map.size()).whenComplete((result, error) -> {
+    getPrimitive(name).thenCompose(collection -> collection.size()).whenComplete((result, error) -> {
       if (error == null) {
         response.resume(Response.ok(result).build());
       } else {
@@ -133,37 +125,18 @@ public class AtomicMapResource extends PrimitiveResource<AsyncAtomicMap<String, 
     });
   }
 
-  @POST
-  @Path("/{name}/clear")
+  @DELETE
+  @Path("/{name}")
   public void clear(
       @PathParam("name") String name,
       @Suspended AsyncResponse response) {
-    getPrimitive(name).thenCompose(map -> map.clear()).whenComplete((result, error) -> {
+    getPrimitive(name).thenCompose(collection -> collection.clear()).whenComplete((result, error) -> {
       if (error == null) {
-        response.resume(Response.noContent().build());
+        response.resume(Response.ok().build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
       }
     });
-  }
-
-  /**
-   * Versioned JSON result.
-   */
-  static class VersionedResult {
-    private final Versioned<String> value;
-
-    public VersionedResult(Versioned<String> value) {
-      this.value = value;
-    }
-
-    public String getValue() {
-      return value.value();
-    }
-
-    public long getVersion() {
-      return value.version();
-    }
   }
 }
