+++ rest/src/main/java/io/atomix/rest/resources/DistributedLockResource.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.core.lock.impl;
+package io.atomix.rest.resources;
 
-import io.atomix.core.lock.AsyncAtomicLock;
-import io.atomix.core.lock.AtomicLockConfig;
-import io.atomix.core.lock.AtomicLockType;
-import io.atomix.primitive.resource.PrimitiveResource;
+import io.atomix.core.lock.AsyncDistributedLock;
+import io.atomix.core.lock.DistributedLockConfig;
+import io.atomix.core.lock.DistributedLockType;
+import io.atomix.rest.AtomixResource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,12 +35,13 @@ import javax.ws.rs.core.Response;
 /**
  * Distributed lock resource.
  */
-@Path("/atomic-lock")
-public class AtomicLockResource extends PrimitiveResource<AsyncAtomicLock, AtomicLockConfig> {
-  private static final Logger LOGGER = LoggerFactory.getLogger(AtomicLockResource.class);
+@AtomixResource
+@Path("/lock")
+public class DistributedLockResource extends PrimitiveResource<AsyncDistributedLock, DistributedLockConfig> {
+  private static final Logger LOGGER = LoggerFactory.getLogger(DistributedLockResource.class);
 
-  public AtomicLockResource() {
-    super(AtomicLockType.instance());
+  public DistributedLockResource() {
+    super(DistributedLockType.instance());
   }
 
   @POST
@@ -49,7 +50,7 @@ public class AtomicLockResource extends PrimitiveResource<AsyncAtomicLock, Atomi
   public void lock(@PathParam("name") String name, @Suspended AsyncResponse response) {
     getPrimitive(name).thenCompose(lock -> lock.lock()).whenComplete((result, error) -> {
       if (error == null) {
-        response.resume(Response.ok(result.value()).build());
+        response.resume(Response.ok().build());
       } else {
         LOGGER.warn("{}", error);
         response.resume(Response.serverError().build());
