+++ rest/src/test/java/io/atomix/rest/impl/EventLogTest.java
@@ -15,7 +15,6 @@
  */
 package io.atomix.rest.impl;
 
-import io.atomix.core.utils.EventLog;
 import org.junit.Test;
 
 import java.util.concurrent.CompletableFuture;
